==> 03_GA_Senate_Runoff.Rmd <==
---
title: "TV ads in Georgia Senate Runoffs"
author: "Aaron Kessler, Data/Investigations"
date: "`r format(Sys.time(), '%B %d, %Y')`"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(tidyverse)
library(lubridate)
library(janitor)
library(scales)
library(gt)
library(kableExtra)
options(dplyr.summarise.inform = FALSE)

#to pull data directly from AA mysql db, run this line, otherwise comment out:
source("00_b_load_data_pullfrom_db_POSTELECTION.R")


#load saved ADMO files from db pulls done in step 00_B
senate_ads_all <- readRDS("raw_data/senate_ads_all_postelex.rds")
house_ads_all <- readRDS("raw_data/house_ads_all_postelex.rds")
prez_ads_all <- readRDS("raw_data/prez_ads_all_postelex.rds")

#add field to mark senate/house
senate_ads_all <- senate_ads_all %>% 
  mutate(
    office = "senate"
  )

house_ads_all <- house_ads_all %>% 
  mutate(
    office = "house"
  )

prez_ads_all <- prez_ads_all %>% 
  mutate(
    office = "prez"
  )


# combine into one (can also work with each separately if desired)
admo_main <- bind_rows(prez_ads_all, senate_ads_all, house_ads_all)

#add date-derived fields
admo_main <- admo_main %>% 
  mutate(
    airdate_year = year(airdate),
    airdate_month = month(airdate),
    airdate_isoweek = isoweek(airdate),
    airdate_day = day(airdate)
  )

#pull out GA runoff records only, tv ads
ga_runoff_tvads <- admo_main %>% 
  filter(
    # election_state == "GA",
    election %in% c("GA Senate 2020 General Runoff", "GA Senate Special 2020 General Runoff"),
    !station %in% c("Facebook", "Google")) #take out social media ads so it's tv airings only


#### DELTA - SPENDING CONTRACTS ####

#load saved delta spending file
admo_delta_postelex <- readRDS("raw_data/admo_delta_postelex.rds")

admo_delta_postelex <- admo_delta_postelex %>% 
  mutate(
    spend_date_year = year(spend_date),
    spend_date_month = month(spend_date),
    spend_date_isoweek = isoweek(spend_date),
    spend_date_day = day(spend_date)
  )

ga_runoff_delta <- admo_delta_postelex %>% 
  filter(election_state == "GA",
         election %in% c("GA Senate 2020 General Runoff", "GA Senate Special 2020 General Runoff"),
         media_type %in% c("Broadcast", "Cable", "Satellite"), #limit to just TV
         !station %in% c("Facebook", "Google")) #should be filtered by prev line but just in case



# #TO FOCUS JUST ON ONE SPECIFIC DAY
# ga_runoff_tvads <- ga_runoff_tvads %>% 
#   filter(airdate == "2020-11-26")
# 
# ga_runoff_delta <- ga_runoff_delta %>% 
#   filter(spend_date == "2020-11-26")


```

####
The following summaries are compiled from Ad Analytics' database of all political ads that have aired in U.S. markets (including cable and satellite providers, along with a sm). Here we're looking only at the ads associated with the two Georgia Senate runoffs that could determine the balance of power in the U.S. Senate. The runoff elections will take place on Jan. 5.  
  
The numbers below reflect the currently available data as of **`r format(Sys.time(), '%I:%M %p on %b. %d')`**.


####
### *Totals*

```{r, echo=FALSE}

ga_runoff_tvads %>% 
  group_by(advertiser_party) %>% 
  summarise(ads_aired = n()) %>%
  arrange(desc(ads_aired)) %>% 
  gt() %>%
  # tab_options(table.align = "left") %>%
  cols_align(
    align = "right",
    columns = vars(ads_aired)
  ) %>% 
  fmt_number(
    columns = vars(ads_aired),
    decimals = 0,
    use_seps = TRUE
  ) %>% 
  # fmt_currency(
  #   columns = vars(est_spent),
  #   decimals = 0,
  #   currency = "USD"
  # ) %>% 
   tab_header(
    title = md("Ads Aired by **Party**"),
    subtitle = "(From Nov. 4th)"
  )

```

####
```{r, echo=FALSE}

ga_runoff_tvads %>% 
  group_by(election, advertiser_party, advertiser) %>% 
  summarise(ads_aired = n()) %>%
  arrange(desc(ads_aired)) %>% 
  gt(rowname_col = "election", groupname_col = "advertiser_party") %>%
  # tab_options(table.align = "left") %>% 
  cols_align(
    align = "right",
    columns = vars(ads_aired)
  ) %>% 
  fmt_number(
    columns = vars(ads_aired),
    decimals = 0,
    use_seps = TRUE
  ) %>% 
  # fmt_currency(
  #   columns = vars(est_spent),
  #   decimals = 0,
  #   currency = "USD"
  # ) %>% 
   tab_header(
    title = md("Ads Aired by **Candidate/Group**"),
    subtitle = "(From Nov. 4th)"
  )

```


####
### *Daily Totals*
```{r, echo=FALSE}

dailycount_byparty <- ga_runoff_tvads %>% 
  count(airdate, advertiser_party, name = "ad_count") # %>% filter(airdate < "2020-11-23")

#area chart by party
ggplot(dailycount_byparty, aes(x = airdate, y = ad_count)) + 
  geom_col(aes(color = advertiser_party, fill = advertiser_party), 
            alpha = 0.5, position = position_dodge(preserve = "single")) +
  scale_color_manual(values = c("#00AFBB", "#F2785D")) +
  scale_fill_manual(values = c("#00AFBB", "#F2785D")) +
  labs(title = "Daily Count of Ad Spots, By Party", 
       subtitle = "",
       x = "",
       y = "") +
  theme_minimal() +
  theme(plot.title = element_text(hjust = 0.5)) +
  scale_y_continuous(labels = comma)



```

####
```{r, echo=FALSE}

dailycount_bycand <- ga_runoff_tvads %>% 
  count(airdate, advertiser, name = "ad_count")

#bar chart for CANDIDATES
ggplot(dailycount_bycand, aes(x = airdate, y = ad_count)) + 
  geom_col(aes(color = advertiser, fill = advertiser), 
           alpha = 0.5, position = position_dodge(preserve = "single")) +
  # scale_color_manual(values = c("#DC3A3A", "#4BAFD8", "#E8857E", "#2E90B8")) +
  # scale_fill_manual(values = c("#DC3A3A", "#4BAFD8", "#E8857E", "#2E90B8")) +
  labs(title = "Daily Count of Ad Spots, By Candidate", 
       subtitle = "",
       x = "",
       y = "") +
  theme_minimal() +
  theme(plot.title = element_text(hjust = 0.5)) +
  scale_y_continuous(labels = comma)


```




##
### *Tone*
```{r, echo=FALSE}
ga_runoff_tvads %>% 
  count(advertiser_party, tone, name = "ads_aired") %>% 
  gt() %>%
  # tab_options(table.align = "left") %>% 
  cols_align(
    align = "right",
    columns = vars(ads_aired)
  ) %>% 
  fmt_number(
    columns = vars(ads_aired),
    decimals = 0,
    use_seps = TRUE
  ) %>% 
   tab_header(
    title = md("Ad Tone by **Party**"),
    subtitle = "(From Nov. 4th)"
  )



```

####

```{r, echo=FALSE}

ga_runoff_tvads %>% 
  group_by(advertiser_party, advertiser, tone) %>% 
  summarise(ads_aired = n()) %>%
  gt(rowname_col = "election", groupname_col = "advertiser_party") %>%
  # tab_options(table.align = "left") %>% 
  cols_align(
    align = "right",
    columns = vars(ads_aired)
  ) %>% 
  fmt_number(
    columns = vars(ads_aired),
    decimals = 0,
    use_seps = TRUE
  ) %>% 
  # fmt_currency(
  #   columns = vars(est_spent),
  #   decimals = 0,
  #   currency = "USD"
  # ) %>% 
   tab_header(
    title = md("Ad Tone by **Candidate/Group**"),
    subtitle = "(From Nov. 4th)"
  )

```



##
### *Markets*
```{r, echo=FALSE}

ga_runoff_tvads %>% 
  group_by(market) %>% 
  summarise(ads_aired = n()) %>%
  arrange(desc(ads_aired)) %>% 
  head(10) %>% 
  gt() %>%
  # tab_options(table.align = "left") %>% 
  cols_align(
    align = "right",
    columns = vars(ads_aired)
  ) %>% 
  fmt_number(
    columns = vars(ads_aired),
    decimals = 0,
    use_seps = TRUE
  ) %>% 
  # fmt_currency(
  #   columns = vars(est_spent),
  #   decimals = 0,
  #   currency = "USD"
  # )  %>% 
   tab_header(
    title = md("Top 10 Markets for Ad Airings"),
    subtitle = "(From Nov. 4th)"
  )


```


##
### *Favorite Shows*
```{r, echo=FALSE}

ga_runoff_tvads %>% 
  group_by(program_name) %>% 
  summarise(ads_aired = n()) %>%
  arrange(desc(ads_aired)) %>% 
  head(10) %>% 
  gt() %>%
  # tab_options(table.align = "left") %>% 
  cols_align(
    align = "right",
    columns = vars(ads_aired)
  ) %>% 
  fmt_number(
    columns = vars(ads_aired),
    decimals = 0,
    use_seps = TRUE
  ) %>% 
  # fmt_currency(
  #   columns = vars(est_spent),
  #   decimals = 0,
  #   currency = "USD"
  # )  %>% 
   tab_header(
    title = md("Top Programs for Ad Airings"),
    subtitle = "(From Nov. 4th)"
  )


```


##
# Ad Spending

Ad Analytics also captures data on actual and future ad spending taken from the contracts signed with TV stations. This dataset does not provide the detailed information on airtimes and the nature of the ads themselves, but does allow us to see how much money has been spent on TV advertising so far, as well as the ad buys that have been booked for future days/weeks.


### *Spending to Date*

The numbers below reflect the currently available data on **airtime through today, `r format(Sys.time(), '%b. %d')`**.  

####
```{r, echo=FALSE}

ga_runoff_delta %>% 
  filter(spend_date <= today()) %>% 
  group_by(advertiser_party) %>% 
  summarise(spent = sum(amount)) %>%
  arrange(desc(spent)) %>% 
  gt() %>%
  # tab_options(table.align = "left") %>%
  cols_align(
    align = "right",
    columns = vars(spent)
  ) %>% 
  fmt_currency(
    columns = vars(spent),
    decimals = 0,
    currency = "USD"
  ) %>% 
   tab_header(
    title = md("Ad Spending by **Party**"),
    subtitle = "(through today)"
  )


```


####
```{r, echo=FALSE}

dailyspend_byparty <- ga_runoff_delta %>% 
  filter(spend_date <= today(),
         advertiser_party %in% c("Democrat", "Republican")) %>% 
  group_by(spend_date, advertiser_party) %>% 
  summarise(spent = sum(amount)) %>% 
  ungroup()

#daily chart by party
ggplot(dailyspend_byparty, aes(x = spend_date, y = spent)) + 
  geom_col(aes(color = advertiser_party, fill = advertiser_party), 
            alpha = 0.5, position = position_dodge(preserve = "single")) +
  scale_color_manual(values = c("#00AFBB", "#F2785D")) +
  scale_fill_manual(values = c("#00AFBB", "#F2785D")) +
  labs(title = "Daily Spending on Ads Aired, By Party", 
       subtitle = "",
       x = "",
       y = "") +
  theme_minimal() +
  theme(plot.title = element_text(hjust = 0.5)) +
  scale_y_continuous(labels = dollar)



```



####
```{r, echo=FALSE}

ga_runoff_delta %>% 
  filter(spend_date <= today()) %>% 
  group_by(election, advertiser_party, advertiser) %>% 
  summarise(spent = sum(amount)) %>%
  arrange(desc(spent)) %>% 
  gt(rowname_col = "election", groupname_col = "advertiser_party") %>%
  # tab_options(table.align = "left") %>% 
  cols_align(
    align = "right",
    columns = vars(spent)
  ) %>% 
  fmt_currency(
    columns = vars(spent),
    decimals = 0,
    currency = "USD"
  ) %>% 
   tab_header(
    title = md("Ad Spending by **Candidate/Group**"),
    subtitle = "(through today)"
  )


```



### *Spending TODAY itself*

The numbers below reflect the currently available data on **airtime bought for today, `r format(Sys.time(), '%b. %d')`**.  

####
```{r, echo=FALSE}

ga_runoff_delta %>% 
  filter(spend_date == today()) %>% 
  group_by(advertiser_party) %>% 
  summarise(spent = sum(amount)) %>%
  arrange(desc(spent)) %>% 
  gt() %>%
  # tab_options(table.align = "left") %>%
  cols_align(
    align = "right",
    columns = vars(spent)
  ) %>% 
  fmt_currency(
    columns = vars(spent),
    decimals = 0,
    currency = "USD"
  ) %>% 
   tab_header(
    title = md("Ad Spending by **Party**"),
    subtitle = "(Today's Buys Only)"
  )


```

####
```{r, echo=FALSE}

ga_runoff_delta %>% 
  filter(spend_date == today()) %>% 
  group_by(election, advertiser_party, advertiser) %>% 
  summarise(spent = sum(amount)) %>%
  arrange(desc(spent)) %>% 
  gt(rowname_col = "election", groupname_col = "advertiser_party") %>%
  # tab_options(table.align = "left") %>% 
  cols_align(
    align = "right",
    columns = vars(spent)
  ) %>% 
  fmt_currency(
    columns = vars(spent),
    decimals = 0,
    currency = "USD"
  ) %>% 
   tab_header(
    title = md("Ad Spending by **Candidate/Group**"),
    subtitle = "(Today's Buys Only)"
  )


```




### *Future Spending Booked*

The numbers below reflect the currently available data on **airtime booked for after `r format(Sys.time(), '%b. %d')`**.  
  
(Note that these are *booking contracts only* and may change, and they frequently do in fast-moving campaigns.)

####
```{r, echo=FALSE}

ga_runoff_delta %>% 
  filter(spend_date > today()) %>% 
  group_by(advertiser_party) %>% 
  summarise(to_be_spent = sum(amount)) %>%
  arrange(desc(to_be_spent)) %>% 
  gt() %>%
  # tab_options(table.align = "left") %>%
  cols_align(
    align = "right",
    columns = vars(to_be_spent)
  ) %>% 
  fmt_currency(
    columns = vars(to_be_spent),
    decimals = 0,
    currency = "USD"
  ) %>% 
   tab_header(
    title = md("Future Spending by **Party**"),
    subtitle = "(Contracts Only)"
  )


```

####
```{r, echo=FALSE}

ga_runoff_delta %>% 
  filter(spend_date > today()) %>% 
  group_by(election, advertiser_party, advertiser) %>% 
  summarise(to_be_spent = sum(amount)) %>%
  arrange(desc(to_be_spent)) %>% 
  gt(rowname_col = "election", groupname_col = "advertiser_party") %>%
  # tab_options(table.align = "left") %>% 
  cols_align(
    align = "right",
    columns = vars(to_be_spent)
  ) %>% 
  fmt_currency(
    columns = vars(to_be_spent),
    decimals = 0,
    currency = "USD"
  ) %>% 
   tab_header(
    title = md("Future Spending by **Candidate/Group**"),
    subtitle = "(Contracts Only)"
  )


```


```{r, echo=FALSE}




```



```{r, echo=FALSE}
## Export files ####
write_csv(ga_runoff_tvads, "output/ga_runoff_tvads.csv")

#export for outside charting tools - switch to wide format
dailycount_bycand %>% 
  pivot_wider(names_from = advertiser, values_from = ad_count) %>% 
  replace(is.na(.), 0) %>% 
  write_csv("output/ga_dailycount_bycand_wide.csv")

dailycount_byparty %>% 
  pivot_wider(names_from = advertiser_party, values_from = ad_count) %>% 
  replace(is.na(.), 0) %>% 
  write_csv("output/ga_dailycount_byparty_wide.csv")



```



```{r, echo=FALSE}




```



```{r, echo=FALSE}




```
